autoconf: fix checks for boz constants
As of gcc 10. gfortran is stricter about where boz constants can be used.
Standard Fortran does not allow them to be used within PARAMETER
statements (as used by the test in fortran.m4), so the various boz tests
in this file fail with gcc 10.

This commit fixes this by using the boz constants within DATA statements
(which are legal) rather than PARAMETER statements.